monitoring/requirements.txt
prometheus-client==0.19.0
slack-sdk==3.26.1
PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.0
grafana-api-client==1.0.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
networkx==3.1
cryptography==41.0.3
docker==6.1.3
kubernetes==28.1.0
pandas==2.0.3
flask==2.3.3
flask-cors==4.0.0 